refactor: NewsCategorybar.tsx cateogrybar/ 위치로 이동
